refactor(eventsource): address review feedback
- Remove ENABLE_REPO_ACTIVITY_EVENTS config flag (not necessary)
- Rename package services/eventsourcenotifier -> services/eventsource
- Rename event type "merged" -> "pr-merged"
- Replace inline styles with CSS classes (repo-activity-banner-* in repo.css)
